requirements.txt
pylint==2.6.0
pytest==6.2.2    
pytest-cov==2.11.1   
pytest-mock==3.5.1
flake8==3.8.4
addict==2.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coverage==5.4
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
python-dotenv==0.15.0
hjson==3.0.2
